binance-auto-trader/requirements.txt
python-binance
pandas
numpy
ta
requests
python-dotenv
